fixup! fs_rom: clear dataspace if lookup failed
Issue gapfruit#1889, related to #23 (see also commit [1])

- [1] 04cf6ea3abfa3bb301dac59e16b003b9d45af214

+++ repos/os/src/server/fs_rom/main.cc
@@ -318,7 +318,7 @@ class Fs_rom::Rom_session_component : public  Rpc_object<Rom_session>
 			catch (Lookup_failed)     {
 				log(_file_path, " ROM file is missing");
 				if (_file_size > 0) {
-					memset(_file_ds.local_addr<char>(), 0x00, _file_size);
+					memset(_file_ds.local_addr<char>(), 0x00, (size_t)_file_size);
 					_file_size = 0;
 					Signal_transmitter(_sigh).submit();
 				}
